fastq files identified with underscore *_R1*
Account for edge case where a sample has "R1" in the identifier and is
paired end so the  input directory contains *R1*_R2.fastq.gz which would
be treated as READ1 by EpitopeID in identtifyEpitope.sh. Still assumes
read1 and read2 data all together in same input directory and
distinguished by an "_R1" and "_R2" in the name. (Underscores to reflect
Illumina convention).