parse hg19 genome to remove haplotype chromosomes
The downloaded hg19 genome contains haplotype chromosomes which we do not want to include in our simulations. This commit adds a parser to remove these chromosomes and updates setup.sh to use it. The code use to generate a new random tag is also commented out so that the random epitope saved to the repo isn't overwritten.