structure log and results directory for simulations
Add logs directory for job submissions to save stderr and stdout to. Also add results directories where simulated datasets and GenoPipe results will be output to. The READMEs serve as descriptors and placeholders. The .gitignore file was also updated to keep the large files generated out of the remote repo