add scripts for depth simulation on CTCF-R500 background
These four PBS submission scripts are formatted similarly with the other depth simulations except with human this time. They continue to increment the seed from 8000 to 11999 to keep seed inputs independent of other simulations. Again, the user will need to replace the working directory before executing.

Note that CPU and memory were bumped for 50M read simulations and the walltime was adjusted to fit different runtimes for human-based simulations.

.gitignore was update to exclude human results (SyntheticEpitope/results/hg*).