generalize organism for run of on EpiID depth sim
This script was generalized to parse the "LOCUS" bash variable to identify the reference organism and stored into "REF." This bash variable is later used to figure out which EpiID db to use for the EpitopeID input.

In addition to this update I added a module load line for samtools which is used by the EpitopeID tool.

I also added a note on walltime for other yeast data read counts.