add PBS script to run EpiID on mix yeast sim
The PBS submission script uses job arrays to create a job for each mix ratio tested in yeast. Once EpitopeID finishes running on all the simulated data, the time to execute is recorded and the results are tallied to determine how often the epitope insertion was successfully detected for each strain background.