pull metadata for ENCODE eGFP samples
This commit includes an update to the README, the metadata file used to identify the accession codes and md5 checksums for the files to download from ENCODE. These files will be used to demonstrate real-world utility of the EpitopeID module. The metadata is downloaded with the `get_metadata.py` script.