add assay type to metadata info collected and remake metadata file
The scripts/get_metadata.py script is modified to add the assay type as a metadata field collected and "HiC" assays are filtered out because they would cause confusion to the EpitopeID results and would not yield very robust results. More comments were also added to the script for better clarity.

The older copy of metadata was removed and the newer metadata added in this commit with appropriate dates.