add files to download FASTQ from ENCODE for EpiID
Add a PBS script for downloading each FASTQ file in a job-array based on the metadata file generated previously. This script also performs a checksum to ensure the file was properly downloaded.

Add placeholding descriptor READMEs for the logs/ and results/ directories.

Update .gitignore file to avoid saving all the large FASTQ files and logs.