add scripts to analyze all raw EpiID output
The python script added here parses both the metadata and raw EpitopeID results in results/ID/*-ID.tab and merges them so we can investigate patterns in the results.

The results from this script are saved to the two text files in results/.

The script for calling the python script with specific arguments is wrapped in the job/02_tally_results.sh bash script.