add 3xFLAG to epitopes checked in ENCODE data
The 3xFLAG sequence obtained from Rick Meyers lab is added to the hg19_EpiID database for the ENCODE runs (not simulation...yet) so the FASTQ files were rerun on the expanded Tag_DB (eGFP, RANDOM_SEQ, and now 3xFLAG). The tallyi
ng script was updated to account for these changes alogn with some minor changes to the column organization. The old results tallies (`results/eGFP_results_analyzed.*`) are replaced with new files (`results/eGFPandFLAG_results_analyzed.*`).

For now, the 3xFLAG sequence is added to input. Eventually add code to setup.sh to setup a second hg19_EpiID that includes this FLAG sequence.