add jobs to run epiID for human 20M
This commit adds new PBS scripts to run EpitopeID on the 1000 simulations for human at 20M reads as job arrays instead of running them in mass. 1000 samples with 20M reads takes too long to run in mass sequntially so this process is parrallelized with job scheduler job arrays.

Since EpitopeID takes an input directory, temp directories are created with the single samples and EpitopeID is run on those temp directories. This allows us to run EpitopeID on a single sample (as opposed to in mass) without adjusting the identify-Epitope.sh script.

This will have to be applied to the 50M human simulations as well.