add jobs to run epiID for human 50M
Similar to previous commit of PBS scripts for running EpitopeID  on human 20M simulations.

This commit adds new PBS scripts to run EpitopeID on the 1000 simulatons for human at 50M reas as job arrays instead of running them in mass. 1000 samples with 50M  reads each takes too long to run in  mass sequntially so this process is parrallelized with job scheduler job arrays.

Since EpitopeID takes an input directory, temp directories are created with the single samples and EpitopeID is run on those temp directories. This allows us to run EpitopeID on a single sample (as opposed to in mass) without adjusting the identify-Epitope.sh script.

The generic mass EpitopeID script (`job/run_EpitopeID_on_depth.pbs`) was also updated to remove the human 50M runs from the mass EpitopeID job array to remove the redundancy and clarify that the human 50M simulations were  run on the parallelized single EpitopeID method.