update DeletionID for Python3
This commit updates the python script used by DeletionID to run with Python3.

The Python2 version of this script relied on the allowance of comparing numeric numpy type to strings so sorting lists of a mix of these two types did not throw an error. I adjusted the script to sort with string types interpreted as a numeric 0. Strings are reported separately so their position in the sort order is unimportant and the purpose of this was just to avoid the thrown exception.

Python 2.7 is no longer supported so publishing a tool using an out-of-date version is not ideal and is inconvenient for potential users.