update DelID utility scripts to Python3
The `calculate_region_coverage.py` was mainly updated to change the print syntax and to catch the exception that is thrown with a `sys.exit()` statement in Python3.

The listed dependency requirements were changed from Python2 to Python3 in the shell scripts.

Python calls in these shell scripts were changed from `python2` to `python`.