add file listing verified KO samples from Puddu
The information from Puddu et al can be used as a "true" set for evaluating the performance of DeletionID. They used an alternative approach for verifying knockouts so this can be an orthogonal verification of DeletionID's performance.

README is updated to describe the new file addition and fix a small typo.