Add scripts to setup StrainID simulations
The scripts for creating a synthetic genome with the variants are added with this commit. This includes setting up a symlink to the VCF files in the `paper/db` directory in the setup.sh script, a PBS wrapper script for recording specific calls of the `scripts/add_VCF_into_Genomic_FASTA.py` with the specific strain inputs, and updates to the `.gitignore` file to include the synthetic_genomes directory that houses the genomes to simulate from.