remove uncommon strain profiles
RedStar is an uncommon strain to use in genomics and as the largest VCF, it significantly contributes to runtime so this file was stripped to make the default yeast run of StrainID more efficient.

HEKTE has so few variants that in such a large genome as hg19, coverage may be insufficient for most human runs so this VCF was stripped from the default run of humans.