update generate_synthetic_genomes
The PBS script for generating the synthetic genomes to simulate off of is updated in this commit to point to the correct sacCer3 VCF files that contain ALL variants, and not just the subset of variants unique to the strain. The change in reference VCF files for the sacCer3 default run in a prior update precipitated the need for this update.