Add ecoli reference files for EpitopeID
This commit includes the reference files for running EpitopeID on E. coli data using the W3110 strain reference assembly. The utility scripts are also updated to allow users to rebuild the annotation files with different bin sizes. The .gitignore file was updated to exclude the BWA index files but the ecoli genome (small file size) is included in this commit.