update docusaurus version
This commit updates the docusaurus version from the alpha to beta version. This includes restructuring the docusaurus configs and updating image links within the markdown to use the new image linking syntax. Missing image for Figure 1 was added, unused docusaurus image defaults were removed, and package updates are also included in this commit