update docusaurus Homepage
The homepage is updated in this commit to match the default initialization format currently set up on the master branch. CSS and homepage (index.js) files are broken up to be more modular and this commit updates the broken up files to match the content of the old docusaurus version.