update to skip bp-count of genomic FASTA
Simulations take some time to run and a large chunk of time is spent by the generate_random_BED_from_Genomic_FASTA.pl script looping through each bp of the genome to count up the genome size. This commit includes an alternative to bypass most of that time spent by creating a FASTA index `.fai` file using samtools for each synthetic genome and creating a variation of the perl script that parses a `.fai` file instead.