adjust SRA download scripts to reformat FASTQ
The FASTQ format from SRA downloads, while valid, does not work with EpitopeID. Be sure to note this in the documentation. This script was adjusted to rename the files with *_R1.fastq.gz and *_R2.fastq.gz instead of *_1.fastq.gz and *_2.fastq.gz. The files themselves were also processed with sed to strip out the header descriptions for the quality scores (plain "+" on those lines) and to replace the SRR read id with the next token (raw Illumina-style read id).

Minor changes: rename conda environment and sra-tools fastq dump used to retrieve FQ instead of parallel-fastq-dump